WIP: Debug branch for UIT testing
Very much work in progress, do not merge. This works around some changes
in response to password/challenge response in authentication, and once
authenticated creates a log of calls as a test job is submitted. The
available hosts could be retrieved, the first directory is created, but
there is never a response to the directory creation, and so submission
is halted at that point.